Remove unused transforms.
Currently encoder is 13.1 kLOC, benchmark results:

```
Encoding                           kPixels    Bytes          BPP  E MP/s  D MP/s     Max norm        pnorm       BPP*pnorm   Bugs
---------------------------------------------------------------------------------------------------------------------------------
custom:jxl:cjxl_tiny:djxl:-d1        13270  2683621    1.6178027   2.341   4.363   1.77444887   0.61091871  0.988345933368      0
custom:jxl:cjxl_tiny:djxl:-d4        13270   987668    0.5954089   2.589   3.280   4.62301445   1.56728331  0.933174497963      0
```
