--- encoder/chroma_from_luma.cc
@@ -1,22 +1,0 @@
-// Copyright (c) the JPEG XL Project Authors.
-//
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file or at
-// https://developers.google.com/open-source/licenses/bsd
-
-#include "encoder/chroma_from_luma.h"
-
-namespace jxl {
-
-ColorCorrelationMap::ColorCorrelationMap(size_t xsize, size_t ysize, bool XYB)
-    : ytox_map(DivCeil(xsize, kColorTileDim), DivCeil(ysize, kColorTileDim)),
-      ytob_map(DivCeil(xsize, kColorTileDim), DivCeil(ysize, kColorTileDim)) {
-  ZeroFillImage(&ytox_map);
-  ZeroFillImage(&ytob_map);
-  if (!XYB) {
-    base_correlation_b_ = 0;
-  }
-  RecomputeDCFactors();
-}
-
-}  // namespace jxl
